refactor input batch
