+++ vllm_spyre/v1/worker/spyre_model_runner.py
@@ -16,8 +16,8 @@ from vllm.v1.outputs import SamplerOutput
 
 from vllm_spyre.model_executor.model_loader.spyre import SpyreCausalLM
 from vllm_spyre.platform import SpyrePlatform
-from vllm_spyre.v1.worker.spyre_input_batch import (CachedRequestState,
-                                                    InputBatch)
+from vllm_spyre.v1.worker.spyre_input_batch import (SamplingInputBatch,
+                                                    SamplingRequestState)
 
 if TYPE_CHECKING:
     from vllm.v1.core.sched.output import (CachedRequestData, NewRequestData,
@@ -96,7 +96,7 @@ class SpyreModelRunner:
         self.warmup_mode = True
 
         # Batch state
-        self.input_batch = InputBatch(
+        self.input_batch = SamplingInputBatch(
             max_num_reqs=vllm_config.scheduler_config.max_num_seqs,
             max_model_len=vllm_config.model_config.max_model_len,
             device=self.device,
@@ -306,7 +306,7 @@ class StaticBatchingSpyreModelRunner(SpyreModelRunner):
             else:
                 generator = None
 
-            req_state = CachedRequestState(
+            req_state = SamplingRequestState(
                 req_id=req_id,
                 prompt_token_ids=request_data.prompt_token_ids,
                 sampling_params=sampling_params,
@@ -319,7 +319,7 @@ class StaticBatchingSpyreModelRunner(SpyreModelRunner):
         self.input_batch.padded_batch_size = padded_batch_size
 
         # Refresh sampling metadata after all request are added to the batch
-        self.input_batch.refresh_sampling_metadata()
+        self.input_batch.refresh()
 
         # padding to compiled batch size
         while len(input_token_list) < padded_batch_size:
@@ -419,7 +419,7 @@ class StaticBatchingSpyreModelRunner(SpyreModelRunner):
             if scheduler_output.finished_req_ids:
                 for req_id in scheduler_output.finished_req_ids:
                     self.input_batch.remove_request(req_id)
-                self.input_batch.refresh_sampling_metadata()
+                self.input_batch.refresh()
 
             return self._prepare_decode(scheduler_output.scheduled_cached_reqs)
 
@@ -588,7 +588,7 @@ class ContinuousBatchingSpyreModelRunner(SpyreModelRunner):
 
         # TODO: Remove this once we can prefill and decode
         # in the same step
-        self.prefill_batch = InputBatch(
+        self.prefill_batch = SamplingInputBatch(
             # TODO: review this, currently we only support prefill for
             # `batch_size=1`
             max_num_reqs=1,
@@ -669,7 +669,7 @@ class ContinuousBatchingSpyreModelRunner(SpyreModelRunner):
             else:
                 generator = None
 
-            req_state = CachedRequestState(
+            req_state = SamplingRequestState(
                 req_id=req_id,
                 prompt_token_ids=request_data.prompt_token_ids,
                 sampling_params=sampling_params,
@@ -681,8 +681,8 @@ class ContinuousBatchingSpyreModelRunner(SpyreModelRunner):
             self.prefill_batch.add_request(req_state)
 
         # Refresh sampling metadata after all request are added to the batch
-        self.input_batch.refresh_sampling_metadata()
-        self.prefill_batch.refresh_sampling_metadata()
+        self.input_batch.refresh()
+        self.prefill_batch.refresh()
 
         # TODO: Review this in the future
         # prefills are always of batch size 1 for this milestone
